- Added CSRF to storages - Added method to get storages of a the current App. - Remove $public for initStorage in getFile() method. We only need to use this method for private storage as public storage will have a public link. Note: This needs more logic work. - Added error handling for no storage configured for the application.

+++ applications/Dash/Components/Storages/StoragesComponent.php
@@ -148,6 +148,10 @@ class StoragesComponent extends BaseComponent
      */
     public function removeAction()
     {
+        if (!$this->checkCSRF()) {
+            return;
+        }
+
         if ($this->request->isPost() && isset($this->postData()['uuid'])) {
 
             $this->storages->removeFile($this->postData()['uuid']);
